:recycle:  fix circular dependencies in types package
