+++ packages/types/src/index.d.ts
@@ -1,29 +1,3 @@
-export type {
-  GraphQLArgument,
-  GraphQLDirective,
-  GraphQLEnumType,
-  GraphQLField,
-  GraphQLFieldMap,
-  GraphQLInputFieldMap,
-  GraphQLInputObjectType,
-  GraphQLInterfaceType,
-  GraphQLNamedType,
-  GraphQLObjectType,
-  GraphQLScalarType,
-  GraphQLSchema,
-  GraphQLSchemaConfig,
-  GraphQLType,
-  GraphQLUnionType,
-} from "graphql/type";
-
-export type { LoadSchemaOptions } from "@graphql-tools/load";
-
-export type { GraphQLExtensionDeclaration } from "graphql-config";
-
-export type { ASTNode, DirectiveNode } from "graphql/language";
-
-export type Maybe<T> = T | null | undefined;
-
 // @graphql-markdown/core
 export * from "./core";
